Advanced/blog_full_project/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cssselect==1.2.0
cycler==0.11.0
d==0.2.2
exceptiongroup==1.1.3
Flask==2.3.3
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.42.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
multidict==6.0.4
numpy==1.25.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pefile==2023.2.7
Pillow==10.0.0
pycparser==2.21
Pygments==2.16.1
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PyJWT==2.8.0
pyparsing==3.0.9
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.2
redis==5.0.0
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spotipy==2.23.0
SQLAlchemy==2.0.20
trio==0.22.2
trio-websocket==0.10.3
twilio==8.7.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.2
